[refactor] move searchbar component to search page path, remove unnecessary code
