+++ my-code/src/components/pages/Search/Search.js
@@ -1,10 +1,10 @@
 import React, { useState } from 'react';
+import SearchBar from './SearchBar';
 import EmptyState from './EmptyState';
 import MovieCard from './MovieCard';
 import MovieCardSkeleton from './MovieCardSkeleton';
 import Layout from '../../shared/Layout';
 import NotFound from '../../shared/NotFound';
-import SearchBar from '../../shared/SearchBar';
 import useFetch from '../../../hooks/useFetch';
 import styles from './Search.css';
